requirements.txt
virtualenv==20.4.6
azure-storage-blob==12.8.1
Flask==2.0.0
nltk==3.6.2
pymongo==3.11.4
python-docx==0.8.11
pdfminer
azure-storage







